Refactored helper functions into utils module.
