+++ tests/test_api_delete.py
@@ -20,7 +20,7 @@ import logging
 import os
 
 from django.test import TestCase
-from django_drf_filepond.views import _get_file_id
+from django_drf_filepond.utils import _get_file_id
 from django_drf_filepond.models import StoredUpload
 
 import django_drf_filepond.drf_filepond_settings as local_settings
