+++ tests/test_api_remote.py
@@ -16,7 +16,7 @@ import logging
 import os
 
 from django.test import TestCase
-from django_drf_filepond.views import _get_file_id
+from django_drf_filepond.utils import _get_file_id
 from django.core.files.uploadedfile import SimpleUploadedFile
 from django_drf_filepond.models import TemporaryUpload, StoredUpload
