+++ tests/test_load_view.py
@@ -1,7 +1,7 @@
 import logging
 
 from django.test.testcases import TestCase
-from django_drf_filepond.views import _get_file_id
+from django_drf_filepond.utils import _get_file_id
 from django_drf_filepond.models import TemporaryUpload, StoredUpload
 from django.core.files.uploadedfile import SimpleUploadedFile
 from django.urls import reverse
