+++ tests/test_uploaders_standard.py
@@ -6,7 +6,7 @@ from rest_framework.request import Request
 from django_drf_filepond.uploaders import FilepondStandardFileUploader
 from rest_framework.exceptions import ParseError
 from django.core.files.uploadedfile import InMemoryUploadedFile
-from django_drf_filepond.views import _get_file_id
+from django_drf_filepond.utils import _get_file_id
 from django.contrib.auth.models import AnonymousUser
 from django_drf_filepond.models import TemporaryUpload
 from django_drf_filepond.renderers import PlainTextRenderer
